EXP-9/backend/requirements.txt
django==4.2.1
django-rest-framework==0.1.0  
requests==2.2.0